Clean up on main file
