+++ lib/main.dart
@@ -8,23 +8,13 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:hive_flutter/hive_flutter.dart';
-import 'package:seeds/blocs/authentication/viewmodels/bloc.dart';
-import 'package:seeds/blocs/deeplink/viewmodels/deeplink_bloc.dart';
-import 'package:seeds/blocs/deeplink/viewmodels/deeplink_state.dart';
 import 'package:seeds/datasource/local/member_model_cache_item.dart';
 import 'package:seeds/datasource/local/models/vote_model_adapter.dart';
 import 'package:seeds/datasource/local/settings_storage.dart';
 import 'package:seeds/datasource/remote/firebase/firebase_push_notification_service.dart';
 import 'package:seeds/datasource/remote/firebase/firebase_remote_config.dart';
 import 'package:seeds/domain-shared/bloc_observer.dart';
-import 'package:seeds/navigation/navigation_service.dart';
-import 'package:seeds/screens/app/app.dart';
-import 'package:seeds/screens/authentication/login_screen.dart';
-import 'package:seeds/screens/authentication/sign_up/signup_screen.dart';
-import 'package:seeds/screens/authentication/splash_screen.dart';
-import 'package:seeds/screens/authentication/verification/verification_screen.dart';
-import 'package:seeds/screens/onboarding/onboarding_screen.dart';
-import 'package:seeds/seeds_material_app.dart';
+import 'package:seeds/seeds_app.dart';
 
 bool get isInDebugMode {
   var inDebugMode = false;
@@ -78,52 +68,4 @@ Future<void> main(List<String> args) async {
       });
     }
   });
-}
-
-class MainScreen extends StatelessWidget {
-  const MainScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<AuthenticationBloc, AuthenticationState>(
-      builder: (context, state) {
-        final navigationService = NavigationService.of(context);
-        switch (state.authStatus) {
-          case AuthStatus.emptyAccount:
-          case AuthStatus.recoveryMode:
-            return BlocBuilder<DeeplinkBloc, DeeplinkState>(
-              builder: (context, deepLinkState) {
-                if (deepLinkState.inviteLinkData != null) {
-                  return SeedsMaterialApp(
-                    home: SignupScreen(deepLinkState.inviteLinkData!.mnemonic),
-                  );
-                } else {
-                  return SeedsMaterialApp(
-                    home: state.authStatus == AuthStatus.emptyAccount
-                        ? const OnboardingScreen()
-                        : SeedsMaterialApp(
-                            home: const LoginScreen(),
-                          ),
-                    navigatorKey: navigationService.onboardingNavigatorKey,
-                    onGenerateRoute: navigationService.onGenerateRoute,
-                  );
-                }
-              },
-            );
-          case AuthStatus.emptyPasscode:
-            return SeedsMaterialApp(home: const VerificationScreen());
-          case AuthStatus.locked:
-            return SeedsMaterialApp(home: const VerificationScreen());
-          case AuthStatus.unlocked:
-            return SeedsMaterialApp(
-              navigatorKey: navigationService.appNavigatorKey,
-              onGenerateRoute: navigationService.onGenerateRoute,
-              home: const App(),
-            );
-          default:
-            return SeedsMaterialApp(home: const SplashScreen());
-        }
-      },
-    );
-  }
 }
